+++ psrc/t8dev/cli/t8dev/emulator/virtualt.py
@@ -1,7 +1,7 @@
 ' VirtualT TRS-80 Model 100 (and friends) emulator suite. '
 
-from    contextlib  import chdir
 from    shutil  import  copy, copytree
+import  os
 
 from    t8dev.cli.t8dev.emulator.suite  import Suite
 from    t8dev.cli.t8dev.util  import runtool
@@ -18,5 +18,12 @@ class VirtualT(Suite):
         copytree(path.tool('src/virtualt/ROMs/'),
                  self.emudir('ROMs'),
                  dirs_exist_ok=True)
-        with chdir(self.emudir()):
+
+        #   `contextlib.chdir` is not available until Python 3.11.
+        #   If we need to do more of this, we should build or own decorator.
+        oldcwd = os.getcwd()
+        try:
+            os.chdir(self.emudir())
             runtool('virtualt')
+        finally:
+            os.chdir(oldcwd)